src/requirements.txt
Flask
pylint
GitPython
Flask-Mail